+++ CHANGELOG.md
@@ -39,7 +39,8 @@ Change Log
 34. Refined app.coop(e, ...), it now takes arbitrary arguments after e;
 35. Added view.channels option with 'view:channel-hooked' meta event;
 36. Improved view 'ready' timing when given {}/[] empty data; 
-37. Upgraded to jQuery 2.2, rapheal 2.2 and flexlayout 0.3;
+37. Added app.upload() api for direct file upload without editor;
+38. Upgraded to jQuery 2.2, rapheal 2.2 and flexlayout 0.3;
 
 working on ...